+++ packages/cli/tsdown.config.ts
@@ -4,13 +4,22 @@ import { createConfigSchema } from './src/config';
 import { z } from 'zod';
 
 export default defineConfig({
-  entry: ['./src/index.ts', './src/build/index.ts'],
+  entry: [
+    './src/{index,config}.ts',
+    './src/registry/{client,schema}.ts',
+    './src/registry/installer/index.ts',
+    './src/build/index.ts',
+  ],
   format: 'esm',
   dts: true,
   fixedExtension: false,
   target: 'node22',
   deps: {
     onlyAllowBundle: [],
+  },
+  exports: {
+    enabled: true,
+    exclude: ['./index'],
   },
   async onSuccess() {
     console.log('JSON schema generated');