+++ packages/radix-ui/tsdown.config.ts
@@ -17,9 +17,11 @@ export default defineConfig({
   },
   async onSuccess() {
     await compileInline();
-    // wait until https://github.com/rolldown/tsdown/issues/472
+
     let content = (await fs.readFile('dist/components/image-zoom.js')).toString();
-    content = content.replaceAll(`import "./image-zoom2.js";`, `import "./image-zoom.css";`);
+    const lines = content.split('\n');
+    lines.splice(1, 0, `import "./image-zoom.css";`);
+    content = lines.join('\n');
     await fs.writeFile('dist/components/image-zoom.js', content);
     console.log('CSS import updated');
   },