+++ packages/create-app-versions/CHANGELOG.md
@@ -1,5 +1,21 @@
 # create-fumadocs-versions
 
+## 16.0.71
+
+### Patch Changes
+
+- Updated dependencies [054da73]
+  - fumadocs-core@16.6.16
+  - fumadocs-ui@16.6.16
+
+## 16.0.70
+
+### Patch Changes
+
+- Updated dependencies [86d3abb]
+  - fumadocs-ui@16.6.15
+  - fumadocs-core@16.6.15
+
 ## 16.0.69
 
 ### Patch Changes